requirements.txt
tkinter
cryptography>=3.4.8
pillow>=8.3.2
qrcode>=7.3.1
pyinstaller>=5.0.0
